+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Studio Changelog
 
+## 4.5.17
+
+Release Date: 2026-08-09
+
+### Fixed
+- Normalized the shared Top Bar across Studio/Livex so it is positioned higher and is more transparent.
+- Scaled up the Top Bar title font size slightly and centered it.
+- Removed all custom and shared back buttons associated with page headers and top bars from the entire application.
+
 ## 4.5.16
 
 Release Date: 2026-08-09